Reorder types in union types for consistency
Types in PHP union types have been reordered to match project-wide use for consistency. The change doesn't affect functionality but adheres to codebase formatting standards. This change has been applied to catch blocks, function return types, and function parameters across multiple files. A whitespace error in a loop was also fixed.

+++ cloud-admin/RedLock/Options/RedLockOption.php
@@ -16,9 +16,9 @@ class RedLockOption
     /**
      * 单位ms.
      */
-    public int|float $singleNodesTimeout = 10 * 1000;
+    public float|int $singleNodesTimeout = 10 * 1000;
 
-    public int|float $expireDuration = 10 * 1000;
+    public float|int $expireDuration = 10 * 1000;
 
     /**
      * @return $this
